requirements.txt
langchain
langchain-core
langchain-community

langchain-huggingface

langchain-groq

python-dotenv
streamlit